Fix mongodb user/group on RHEL
Hi, not sure if this was intentional or not. But with mongodb user/group set to be "mongod" on RHEL, the cookbook fails out of the box, looking for 'mongod' user/group that are not there, as well as failing to start mongodb because /var/[lib/log/run]/mongodb are owned by mongodb:mongodb. When I changed these attributes to 'mongodb' the cookbook works out the box. Thanks so much for the work.